% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_helpers.R
\name{add_caption<-}
\alias{add_caption<-}
\title{Append a caption to a string vector of captions}
\usage{
add_caption(x) <- value
}
\arguments{
\item{x}{The list of captions to add to.}

\item{value}{a vector with length 1 or 2.  If \code{length(value) == 1}, it
is the caption text.  If \code{length(value) == 2}, the first element is
the caption text and the second is the label for the figure (to be passed
into \code{latex_label}).}
}
\value{
\code{c(x, paste0(latex_label(value[2]), value[1]))} or \code{c(x, value)}
}
\description{
Append a caption to a string vector of captions
}
