% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hash_to_file.R
\name{add_hash_to_file}
\alias{add_hash_to_file}
\title{Create a string of the filename with its hash as an attribute to allow change detection in drake}
\usage{
add_hash_to_file(filename)
}
\arguments{
\item{filename}{The file to generate a hash for}
}
\value{
The filename with an attribute of `hash` added.
}
\description{
Create a string of the filename with its hash as an attribute to allow change detection in drake
}
\examples{
\dontrun{
my_plan <-
  drake_plan(
    big_filename=
      target(
        command=add_hash_to_file("a/file.xpt"),
        trigger=trigger(condition=TRUE)
      )
  )
}
}
