% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.power.htest.R
\name{as.data.frame.power.htest}
\alias{as.data.frame.power.htest}
\title{Convert a power calculation to a table for reporting.}
\usage{
\method{as.data.frame}{power.htest}(x, ..., digits = 3)
}
\arguments{
\item{x}{A `power.htest` object to convert to a data.frame.}

\item{...}{Ignored}

\item{digits}{Significant digits to report for numeric values.}
}
\description{
Convert a power calculation to a table for reporting.
}
