% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.gg.R
\name{as_gg_list}
\alias{as_gg_list}
\title{Make a gg_list object (a list of ggplots) for knit_printing}
\usage{
as_gg_list(x)
}
\arguments{
\item{x}{The list}
}
\value{
A gg_list object (just adding the class to the list)
}
\description{
Make a gg_list object (a list of ggplots) for knit_printing
}
