% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.tab_list.R
\name{as_tab_list}
\alias{as_tab_list}
\title{Create a tab_list object that enables simpler printing of lists of tables}
\usage{
as_tab_list(x)
}
\arguments{
\item{x}{A data.frame or similar object with at least columns named "table" and "caption"}
}
\value{
`x` with a class of "tab_list" added to its list of classes
}
\description{
Create a tab_list object that enables simpler printing of lists of tables
}
