% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_decision_figure.R
\name{bayesian_decision_figure}
\alias{bayesian_decision_figure}
\title{Make a figure describing the Bayesian decision-making based on a range of lower and upper reference values}
\usage{
bayesian_decision_figure(
  n,
  lrv,
  urv,
  one_sided_prob,
  sd_single_measure,
  sd_scale = c("log", "linear"),
  figure_edge = 0.05,
  figure_steps = 100,
  add_points = NULL
)
}
\arguments{
\item{n}{The number of observations (per group)}

\item{lrv, urv}{The lower and upper reference values (target for a stop or go decision)}

\item{one_sided_prob}{The integrated probability outside of the reference value}

\item{sd_single_measure}{The standard deviation of a single measurement in the assessment}

\item{sd_scale}{Is the standard deviation on the "log" or "linear" scale}

\item{figure_edge}{What integrated probability should be outside of the figure edges?}

\item{figure_steps}{How many points should be shown in the figure?}

\item{add_points}{Should any specific points be added to the figure?}
}
\value{
A list with "data" used to make the figures, "plot_prob", and "plot_lines"
}
\description{
Make a figure describing the Bayesian decision-making based on a range of lower and upper reference values
}
