% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_helpers.R
\name{comma_and}
\alias{comma_and}
\title{Generate a list of text grammatically correctly}
\usage{
comma_and(x, oxford_comma = TRUE, conjunction = "and")
}
\arguments{
\item{x}{a vector to be put into a list.}

\item{oxford_comma}{Should the Oxford comma be used?}

\item{conjunction}{The conjunction to use (without spaces).
\code{conjunction} will usually be "and" or "or".}
}
\value{
A character string of the vector input separated by commas and the
  word "and" as appropriate.
}
\description{
Generate a list of text grammatically correctly
}
