% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonym.R
\name{correct_case}
\alias{correct_case}
\title{Correct the case of a vector to be in a preferred case}
\usage{
correct_case(x, preferred)
}
\arguments{
\item{x}{An object to correct the case of}

\item{preferred}{A character vector of preferred values}
}
\value{
`x` where values that match `tolower(x) == tolower(preferred)` are
  converted to the preferred value.
}
\description{
Correct the case of a vector to be in a preferred case
}
\examples{
correct_case(c("ABC", "Abc", "aBc", "def"), "Abc")
}
\seealso{
Other Text standardization:
\code{\link[=replace_synonym]{replace_synonym()}}
}
\concept{Text standardization}
