% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dated_warning_or_error.R
\name{dated_warning_or_error}
\alias{dated_warning_or_error}
\title{Give a warning or error based on the current date}
\usage{
dated_warning_or_error(date, ...)
}
\arguments{
\item{date}{The date cutoff}

\item{...}{Passed to \code{stop} if todays date is < \code{date} or
\code{warning} otherwise.}
}
\description{
Give a warning or error based on the current date
}
