% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_helpers.R
\name{emax_fun}
\alias{emax_fun}
\alias{inverse_emax}
\title{Compute the Emax or its inverse.}
\usage{
emax_fun(x, e0, emax, ex50, hill = 1)

inverse_emax(effect, e0, emax, ex50, hill = 1)
}
\arguments{
\item{e0}{Baseline effect}

\item{emax}{Maximum effect (at the limit as \code{x} approaches \code{Inf})}

\item{ex50}{Amount of x yielding 50% of the maximum effect}

\item{hill}{The hill slope (also known as the sigmoidicity)}

\item{effect}{The effect}
}
\value{
\code{effect} for \code{emax_fun} or \code{x} for \code{inverse_emax}
}
\description{
Compute the Emax or its inverse.
}
\details{
The equation for the Emax function is

  \deqn{effect = e0 + \frac{emax \times x^{hill}}{ex50^{hill} + x^{hill}}}
}
\section{Functions}{
\itemize{
\item \code{inverse_emax}: Inverse emax
}}

