% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_n.R
\name{expect_n}
\alias{expect_n}
\title{Expect a certain number of values in a logical vector}
\usage{
expect_n(x, n = 1, msg = NULL)
}
\arguments{
\item{x}{The logical vector}

\item{n}{The number of expected values}

\item{msg}{An optional message to display in case of an error}
}
\value{
`x` if `sum(x) == n` and raises an informative error, otherwise
}
\description{
Expect a certain number of values in a logical vector
}
