% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_list.R
\name{filter_list}
\alias{filter_list}
\title{Filter a list of data.frames}
\usage{
filter_list(.data, ..., .ignore_errors = TRUE)
}
\arguments{
\item{.data}{A list of data.frames (or similar)}

\item{...}{Arguments to filter_maybe (eventually passed to dplyr::filter)}

\item{.ignore_errors}{(always true)}
}
\value{
A list of data.frames after filtering-- or NULL
}
\description{
Filter a list of data.frames
}
\examples{
filter_list(list(mtcars, iris), cyl == 6)
}
\seealso{
\code{\link{filter_maybe}}
}
