% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_list.R
\name{filter_maybe}
\alias{filter_maybe}
\title{Filter if it works, otherwise, do nothing}
\usage{
filter_maybe(.data, ..., .ignore_errors = TRUE)
}
\arguments{
\item{.data}{The data to potentially filter}

\item{...}{Arguments (see dplyr::filter)}

\item{.ignore_errors}{(always true)}
}
\value{
Either NULL if an error or the filtered results.
}
\description{
Filter if it works, otherwise, do nothing
}
\seealso{
\code{\link{filter_list}}
}
