% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_less.R
\name{find_less}
\alias{find_less}
\title{Find the choice that is less than a value}
\usage{
find_less(x, choices, include_same = TRUE, none = c("first", "na"))
}
\arguments{
\item{x}{The vector of values to select choices for.}

\item{choices}{The choices that are to be selected from.}

\item{include_same}{Include if the choice is \code{<=} and not strictly
\code{<}.}

\item{none}{What if the \code{x} value is less than (or less than or equal to
if \code{include_same}) all the choices?  Choose the \code{"first"} choice
or insert an \code{NA} value?}
}
\value{
A vector with values from \code{choices} that are less than (or equal
  to) \code{x}
}
\description{
Find the choice that is less than a value
}
\details{
\code{x} and \code{choices} must be comparable by the \code{<} or
  \code{<=} operators.  \code{choices} will be sorted to be in ascending
  order; sorting will remove any \code{NA} values from choices.
}
