% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_no_na.R
\name{fun_no_na}
\alias{fun_no_na}
\alias{max_no_na}
\alias{min_no_na}
\alias{median_no_na}
\alias{mean_no_na}
\alias{sd_no_na}
\title{Perform an operation on a vector that may be empty or all NA with
controlled output.}
\usage{
fun_no_na(x, FUN, zero_len = NULL)

max_no_na(x, zero_len = NA)

min_no_na(x, zero_len = NA)

median_no_na(x, zero_len = NA)

mean_no_na(x, zero_len = NA)

sd_no_na(x, zero_len = NA)
}
\arguments{
\item{x}{The object to perform the operation on.}

\item{FUN}{The function to perform if the object is neither zero length or all \code{NA}}

\item{zero_len}{If \code{x} has length 0, then either return it (if \code{NULL}) or return \code{NA} of the same class as \code{x} (if \code{NA})}
}
\value{
An object of the same class as \code{x}
}
\description{
Perform an operation on a vector that may be empty or all NA with
controlled output.
}
\section{Functions}{
\itemize{
\item \code{max_no_na()}: Maximum

\item \code{min_no_na()}: Minimum

\item \code{median_no_na()}: Median

\item \code{mean_no_na()}: Median

\item \code{sd_no_na()}: Median

}}
\examples{
min_no_na(c(NA, 3))
min_no_na(double(), zero_len=NULL)
max_no_na(c(NA, 3))
}
