% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manage.R
\name{get_data_manage_standard_cols}
\alias{get_data_manage_standard_cols}
\title{Get a vector of column names expected for a dataset.}
\usage{
get_data_manage_standard_cols(data, coltype)
}
\arguments{
\item{data}{A data.frame or similar object}

\item{coltype}{A vector of values to match in the "Column Type" column of
`data` or `NULL` to match all values.}
}
\value{
A vector of the subset of "Column Name" values that are in rows of
  "Column Type".
}
\description{
Get a vector of column names expected for a dataset.
}
\details{
The input `data` must have columns named "Column Type" and "Column
  Name".  The "Column Type" defines arbitrary strings to be matched to subset
  for the "Column Name"s of interest.  The "Column Name" are the names
  themselves.

If not all "Column Type" values are in `data[["Column Type"]]`, an error will
be raised.
}
\seealso{
Other Data Management:
\code{\link[=check_expected_cols]{check_expected_cols()}},
\code{\link[=nonmem_column_order]{nonmem_column_order()}}
}
\concept{Data Management}
