% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmulti.R
\name{ggmulti}
\alias{ggmulti}
\title{Make a list of ggplot2 objects that can be added to and have associated captions}
\usage{
ggmulti(figure, caption)
}
\arguments{
\item{figure}{The ggplot2 object}

\item{caption}{The required caption for the figure}
}
\description{
Make a list of ggplot2 objects that can be added to and have associated captions
}
\examples{
p1 <-
  ggplot2::ggplot(data = data.frame(x = 1:2, y = 2:3), ggplot2::aes(x = x, y = y)) +
  ggplot2::geom_point()
p2 <-
  ggplot2::ggplot(data = data.frame(x = 2:3, y = 4:5), ggplot2::aes(x = x, y = y)) +
  ggplot2::geom_point()
c(
  ggmulti(
    figure = p1,
    caption = "my caption"
  ),
  ggmulti(
    figure = p2,
    caption = "my caption 2"
  )
)
}
