% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_tools.R
\name{grepl_multi_pattern}
\alias{grepl_multi_pattern}
\title{Run grepl on a vector of patterns.}
\usage{
grepl_multi_pattern(pattern, x, ...)
}
\arguments{
\item{pattern}{One or more patterns (see grepl)}

\item{x}{A string vector}

\item{...}{Passed to grepl}
}
\value{
A boolean vector if any pattern is matched
}
\description{
Run grepl on a vector of patterns.
}
\examples{
grepl_multi_pattern(pattern=c("A", "B", "C"), LETTERS)
}
