% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_tools.R
\name{gsub_multi_pattern}
\alias{gsub_multi_pattern}
\title{Substitute from a vector of patterns with a single replacement}
\usage{
gsub_multi_pattern(x, patterns, replacement, ..., verbose = FALSE)
}
\arguments{
\item{x, replacement}{See \code{\link[base]{gsub}}}

\item{patterns}{A vector of patterns as used individually in
\code{\link[base]{gsub}} and \code{\link[base]{grepl}}}

\item{...}{Passed to \code{\link[base]{gsub}} and \code{\link[base]{grepl}}}

\item{verbose}{Signal messages with the count of values that matched each
pattern or no pattern.}
}
\value{
A vector of \code{NA_character_} when no match occurs and the
  replaced value when a match occurs.
}
\description{
Substitute from a vector of patterns with a single replacement
}
\seealso{
\code{\link{grepl_multi_pattern}}
}
