% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_dtc.R
\name{impute_dtc}
\alias{impute_dtc}
\alias{impute_dtc_ntod}
\title{Impute dates and times when data are missing.}
\usage{
impute_dtc(data)

impute_dtc_ntod(data, na_ntod = NA_real_)
}
\arguments{
\item{data}{A data.frame or equivalent object with at least the columns
defined in the details section.}

\item{na_ntod}{What nominal time of day should unscheduled measurements be
imputed as?  (Often \code{0} is selected, but missing is the default.)}
}
\value{
`data` with the columns "ADTC_IMPUTE_METHOD" and "ADTC_IMPUTED"
  added.
}
\description{
Impute dates and times when data are missing.
}
\details{
Dates and times will be imputed based on the following rules:

\itemize{
\item{If both date and time are observed, no the observed value will be used.}
\item{Data are assumed to be grouped by appropriate grouping factors within a
  nominal time so that all times may be at the same time, and data are
  assumed to be sorted in the order specified in the protocol.}
\item{If nominal time since first dose (NTSFD) is missing, no imputation will be
  performed (the measure is assumed to be unscheduled).}
\item{If dates differ within a nominal time measurement, no imputation will be
  performed (a data issue would appear to exist in that case).}
\item{If only one date exists within a nominal time measurement, missing dates
  will be assumed to match the observed date.}
\item{If one or more time exists within a nominal interval, all measurements
  in the interval will be assigned to the median of the times that exist.}
}

Columns used in calculation are:

\itemize{
\item{ADTC: (the date and time) formatted as an ISO8601 datetime without the time
  zone (yyyy-mm-ddThh:mm:ss) where the entire time or the seconds parts are
  optional.}
\item{STUDYID, USUBJID, NTSFD: grouping variables for the study number, subject
  identifier, and nominal time since first dose.}
}
}
\section{Functions}{
\itemize{
\item \code{impute_dtc_ntod()}: imputes based on the typical nominal time of day
(NTOD) for a subject.

}}
\seealso{
Other Imputation:
\code{\link[=impute_time_act_nom]{impute_time_act_nom()}}

Other Date/time imputation:
\code{\link[=impute_time_act_nom]{impute_time_act_nom()}}
}
\concept{Date/time imputation}
\concept{Imputation}
