% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_dtc.R
\name{impute_dtc_helpers}
\alias{impute_dtc_helpers}
\alias{impute_dtc_helper_single_datetime}
\alias{impute_dtc_helper_single_date}
\alias{impute_dtc_helper_median_time}
\alias{impute_dtc_helper_time_ntod}
\title{Impute the date and time within a single study/subject/nominal time where
there is only one date/time combo in the interval.}
\usage{
impute_dtc_helper_single_datetime(date, time, ntime, method)

impute_dtc_helper_single_date(date, time, ntime, method)

impute_dtc_helper_median_time(date, time, ntime, method)

impute_dtc_helper_time_ntod(date, time, ntime, method)
}
\arguments{
\item{date, time}{Date and time parts of the date/time (character)}

\item{ntime}{Nominal time (numeric)}

\item{method}{Existing imputation methods used}
}
\value{
A tibble with columns for date, time, ntime, and method
}
\description{
Impute the date and time within a single study/subject/nominal time where
there is only one date/time combo in the interval.
}
\section{Functions}{
\itemize{
\item \code{impute_dtc_helper_single_datetime()}: Impute for a single date/time within a
scheduled nominal time

\item \code{impute_dtc_helper_single_date()}: Impute for a single date within a scheduled
nominal time

\item \code{impute_dtc_helper_median_time()}: Impute for the median time within a scheduled
nominal time if all happen on the same date

\item \code{impute_dtc_helper_time_ntod()}: Impute for the median time within a scheduled
nominal time of day

}}
