% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_dtc.R
\name{impute_dtc_simplify_time}
\alias{impute_dtc_simplify_time}
\title{Confirm that the value looks like a time, and remove fully unknown times and
zero seconds}
\usage{
impute_dtc_simplify_time(x, drop_zero_seconds = FALSE)
}
\arguments{
\item{x}{A character vector that looks like a time (see details)}

\item{drop_zero_seconds}{Drop :00 for seconds (keeping only hours and
minutes)}
}
\value{
A character vector that is simplified and is a time.  Zero will be
  prepended to single-digit hours.
}
\description{
A time is a 12- or 24-hour time possibly with "UN" as specified by the CDISC
SDTM standard.  It may be of the format "" (empty string), #:##, #:##:##,
##:##, ##:##:##, #:##:UN, ##:##:UN, UN:UN, or UN:UN:UN (where # is a valid
number only allowing hours from 00-23, and "UN" signifies that the value is
unknown).
}
