% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_sd.R
\name{impute_sd}
\alias{impute_sd}
\alias{impute_sd_sd}
\alias{impute_sd_cv}
\alias{impute_sd_se}
\alias{impute_sd_ci}
\alias{impute_sd_iqr}
\alias{impute_sd_range}
\title{Impute standard deviation from measures of dispersion.}
\usage{
impute_sd(point, var1, var2, n, vartype)

impute_sd_sd(point, var1, var2, n, vartype)

impute_sd_cv(point, var1, var2, n, vartype)

impute_sd_se(point, var1, var2, n, vartype)

impute_sd_ci(point, var1, var2, n, vartype)

impute_sd_iqr(point, var1, var2, n, vartype)

impute_sd_range(point, var1, var2, n, vartype)
}
\arguments{
\item{point}{The point estimate (measurement of central tendency)}

\item{var1, var2}{The first and second measures of dispersion (`var2` is not
required for all imputation methods)}

\item{n}{The number of observations contributing to the measurement}

\item{vartype}{The type of dispersion (case insensitive; see Details)}
}
\value{
A vector of imputed standard deviations
}
\description{
Impute standard deviation from measures of dispersion.
}
\details{
\code{impute_sd} selects the imputation method based on
  \code{vartype}.

\code{impute_sd_sd()} imputes the standard deviation from itself
  (not really imputation).  \code{var1} is the standard deviation;
  \code{var2} must be \code{NA}; \code{vartype} must be "SD" (ignoring case);
  and \code{n} is ignored.

\code{impute_sd_cv()} imputes the standard deviation from the
  coefficient of variation (CV).  \code{var1} is the CV; \code{var2} must be
  \code{NA}; \code{vartype} must be "CV", "\%CV", or "CV\%" (ignoring case);
  and \code{n} is ignored.

\code{impute_sd_se()} imputes the standard deviation from the
  standard error (SE).  \code{var1} is the standard error of the mean;
  \code{var2} must be \code{NA}; \code{vartype} must be "SE" or "SEM"
  (ignoring case); and \code{n} is required.  Imputation assumes a sample,
  not population, SE.

\code{impute_sd_ci()} imputes the standard deviation from the
  confidence interval (CI).  \code{var1} is the lower bound of the CI;
  \code{var2} is the upper bound of the CI; \code{vartype} must match the
  regular expression "^([0-9]+)\% CI$" (ignoring case); and \code{n} is
  required.  If only an upper or lower bound of the CI is available, set the
  other to \code{NA}.  Imputation assumes a t-distribution.

\code{impute_sd_iqr()} imputes the standard deviation from the
  inter-quartile range (IQR).  \code{var1} is the lower bound of the IQR;
  \code{var2} is the upper bound of the IQR; \code{vartype} must be "IQR"
  (ignoring case); and \code{n} is required.  Imputation assumes a
  t-distribution.

\code{impute_sd_range()} imputes the standard deviation from the
  range.  \code{var1} is the lower bound of the range; \code{var2} is the
  upper bound of the range; \code{vartype} must be "RANGE" (ignoring case);
  and \code{n} is required.  Imputation is performed by computing \code{(var2
  - var2)/4}.  This should be the last choice of imputation methods.
}
\references{
Cochrane Handbook, version 5.1
  \url{https://handbook-5-1.cochrane.org/chapter_7/7_7_3_data_extraction_for_continuous_outcomes.htm}
}
