% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_dtc.R
\name{impute_time_act_nom}
\alias{impute_time_act_nom}
\title{Impute actual time using nominal time when actual is unavailable}
\usage{
impute_time_act_nom(actual, nominal)
}
\arguments{
\item{actual, nominal}{Actual and nominal times as numeric vectors}
}
\value{
\code{actual} with missing values imputed, as feasible, based on
  nominal times.
}
\description{
Impute actual time using nominal time when actual is unavailable
}
\details{
The rules used are the following, in order.

\itemize{
  \item{If \code{actual} is not missing, use it.}
  \item{If there is at least one non-missing \code{actual} for a given
    \code{nominal}, take the median \code{actual} from the \code{nominal}.}
  \item{If for a missing actual time, the actual time values are in order
    compared to the nominal time:}
  \itemize{
    \item{If both directions with nominal and actual times (before and after)
      are within 24 hours, interpolate between.}
    \item{If only one direction with nominal and actual times (before or
      after) is within 24 hours, extrapolate in that direction ignoring the
      other direction.}
  }
}
}
\seealso{
Other Imputation:
\code{\link[=impute_dtc]{impute_dtc()}}

Other Date/time imputation:
\code{\link[=impute_dtc]{impute_dtc()}}
}
\concept{Date/time imputation}
\concept{Imputation}
