% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaners.R
\name{interesting_cols}
\alias{interesting_cols}
\title{Find the columns that have more than one value}
\usage{
interesting_cols(x, keep_anyway = character(0), ...)
}
\arguments{
\item{x}{The object (typically a data.frame) with columns to check.}

\item{keep_anyway}{A character vector of columns to keep even if they
only have a single value.}

\item{...}{ignored (for now)}
}
\value{
An object of the same class as \code{x} with boring columns
  taken out and the duplicated values placed in an attribute called
  \code{"boring"}.
}
\description{
Find the columns that have more than one value
}
\examples{
interesting_cols(data.frame(A=1:2, B=2))
attr(interesting_cols(data.frame(A=1:2, B=2)), "boring")
}
