% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermingle_lists.R
\name{intermingle_list}
\alias{intermingle_list}
\title{Intermingle elements from a set of lists}
\usage{
intermingle_list(...)
}
\arguments{
\item{...}{lists with the same length as each other.}
}
\value{
A list with the length of the sum of the input lists.
}
\description{
Intermingle elements from a set of lists
}
\examples{
intermingle_list(as.list(1:5), as.list(6:10))
}
