% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_control.R
\name{join_control}
\alias{join_control}
\title{Perform join where the outcome of the join is verifed to match an expected
pattern.}
\usage{
join_control(x, y, join_fun, x_control = "any", y_control = "any", ...)
}
\arguments{
\item{x, y}{tbls to join}

\item{join_fun}{Any function that can combine x and y (called as
`join_fun(x, y, ...)`).  Typically this will be one of `dplyr::left_join`,
`dplyr::right_join`, etc.}

\item{x_control, y_control}{What outcome is expected from the `x`, and `y` tbls? Default is "any" (see details).}

\item{...}{Passed to `join_fun()`}
}
\value{
A joined tbl
}
\description{
Perform join where the outcome of the join is verifed to match an expected
pattern.
}
\details{
Options for `x_control` and `y_control` are below and may be combined:

* `"any"`: Any outcome is acceptable; this overrides all other options.
* `"all"`: Each row from the input must appear in the output at least one
  time.
* `"unique"`: A row may appear in the output zero or one time.
* `"missing"`: At least one row must not match in the new dataset (the values
   must be missing).  This option is rarely used.
* `"nomissing"`: All rows must match in the new dataset (the values must not
  be missing).

The combination of `x_control=c("all", "unique", "nomissing")` (or
`y_control`) is a common need to confirm that all values are present exactly
one time and that there are no missing values.
}
