% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.formula.R
\name{knit_print.formula}
\alias{knit_print.formula}
\title{Generate LaTeX math from an R formula}
\usage{
\method{knit_print}{formula}(x, inline = TRUE, replacements = list(),
  ..., width = 80)
}
\arguments{
\item{x}{A formula}

\item{inline}{Should the formula be generated for use inline or as a separate
line? (In LaTeX notation, with "$" or "$$".)}

\item{replacements}{A named list where whenever the list name is seen as the
text for any complete part of the equation, it is replaced with the value
(which should be a LaTeX equation fragment).}

\item{...}{Currently ignored}

\item{width}{Approximate number of characters for splitting to multiple
lines.  (Not yet implemented.)}
}
\value{
A knitr \code{asis_output()} formula
}
\description{
Generate LaTeX math from an R formula
}
\examples{
knit_print(a~b)
# The same with an equal sign
knit_print(a~b, replacements=list("\\\\sim"="="))
knit_print(a~b/c)
}
