% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.gg.R
\name{knit_print.gg}
\alias{knit_print.gg}
\title{Print a ggplot object with space around it}
\usage{
\method{knit_print}{gg}(x, ..., fig_prefix, fig_suffix)
}
\arguments{
\item{x}{The plot object}

\item{...}{Passed to \code{print}.}

\item{fig_prefix, fig_suffix}{Character strings passed to \code{cat} before
and after printing \code{x} (if not missing).}
}
\value{
\code{x} invisibly
}
\description{
Print a ggplot object with space around it
}
\seealso{
\code{\link{knit_print.gg_list}}
}
