% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.gg.R
\name{knit_print.gg_list}
\alias{knit_print.gg_list}
\title{Print a list of ggplot objects with space around each}
\usage{
\method{knit_print}{gg_list}(x, ..., fig_suffix = "\\\\clearpage\\n")
}
\arguments{
\item{x}{A list of plot objects}
}
\value{
\code{x} invisibly
}
\description{
Print a list of ggplot objects with space around each
}
\seealso{
\code{\link{knit_print.gg}}
}
