% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.sessionInfo.R
\name{knit_print.sessionInfo}
\alias{knit_print.sessionInfo}
\title{Print the sessionInfo in a way that is usable in knitr/rmarkdown reports}
\usage{
\method{knit_print}{sessionInfo}(x, ...)
}
\arguments{
\item{x}{The output of sessionInfo() (if missing, it is generated)}
}
\value{
A knitr \code{asis_output()} version of the session info.
}
\description{
Print the sessionInfo in a way that is usable in knitr/rmarkdown reports
}
