% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.tab_list.R
\name{knit_print.tab_list}
\alias{knit_print.tab_list}
\title{Print a `tab_list` object}
\usage{
\method{knit_print}{tab_list}(
  x,
  ...,
  filename = NULL,
  tab_suffix = "\\n\\n",
  tabfun = pander::pandoc.table.return
)
}
\arguments{
\item{x}{A list of plot objects}

\item{...}{passed to `tabfun`}

\item{filename}{Save the figure to the filename, if provided.  If the
filename contains "%d" (optionally with sprintf-formatting instructions
such as "%03d"), then the filename will be run through
\code{sprintf(filename, seq_along(x))} to generate the filename.}

\item{tab_suffix}{Character strings passed to `cat()` before and after
printing `x` (if not missing).}

\item{tabfun}{The function used to generate a table.  Common examples are
`pander::pandoc.table.return` and `TopicLongTable::topic_long_table`}
}
\value{
The character string to output as a `knitr` `asis_output` object
}
\description{
Print a `tab_list` object
}
