% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.xpose_data}
\alias{logLik.xpose_data}
\alias{logLik.NULL}
\alias{logLik.logical}
\alias{logLik.try-error}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{xpose_data}(object, ...)

\method{logLik}{`NULL`}(object, ...)

\method{logLik}{logical}(object, ...)

\method{logLik}{`try-error`}(object, ...)
}
\arguments{
\item{object}{any object from which a log-likelihood value, or a
   contribution to a log-likelihood value, can be extracted.}

\item{...}{some methods for this generic function require additional
   arguments.}
}
\value{
a logLik object
}
\description{
See \code{stats::logLik()} for use
}
\section{Functions}{
\itemize{
\item \code{logLik(`NULL`)}: logLik for NULL returns an NA object

\item \code{logLik(logical)}: logLik for NA returns an NA object

\item \code{logLik(`try-error`)}: logLik for try-error returns an NA object

}}
\seealso{
\code{\link{NA_logLik_}}
}
