% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaners.R
\name{make_blq}
\alias{make_blq}
\title{Extract numeric, BLQ, and text data from a vector.}
\usage{
make_blq(x, blq_pattern = "^< *\\\\(? *([0-9]*\\\\.?[0-9]+) *\\\\)?$",
  number_pattern = "^(+-)?[0-9]+(\\\\.[0-9]*)?([eE][+-]?[0-9]+)?$",
  replace_blq = 0, ...)
}
\arguments{
\item{x}{The character vector to extract data from.}

\item{blq_pattern}{A regex or vector of regex values for extracting BLQ.  If
a vector, all regexes will be tested.}

\item{number_pattern}{A regex or vector of regex values for finding numbers.}

\item{replace_blq}{The value to place in the "number" column when a BLQ is
matched (typically \code{0} or \code{NA_real_}).}

\item{...}{Parameters passed to \code{grep} and \code{grepl} for
\code{blq_pattern} searching.}
}
\value{
a data.frame with columns named "text", "number", and "blq".
}
\description{
Extract numeric, BLQ, and text data from a vector.
}
\examples{
make_blq(c("1", "A", "<1"))
make_blq(c("1", "A", "<1"), replace_blq=NA_real_)
}
