% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_helpers.R
\name{make_ci}
\alias{make_ci}
\title{Make a confidence interval from a point estimate and standard error.}
\usage{
make_ci(point, se, level = 0.95, transform = NULL, numeric = FALSE)
}
\arguments{
\item{point}{The point estimate (numeric vector)}

\item{se}{The standard error of the estimate (numeric vector)}

\item{level}{The confidence level}

\item{transform}{A function that takes a matrix input and returns a matrix
output of the confidence interval values transformed.}

\item{numeric}{Should the transformed point estimate be returned instead of
the character representation?}
}
\value{
If \code{numeric=FALSE} a character vector of the confidence
  intervals represented as "X [X, X]" where X are numbers with three
  significant figures.  If \code{numeric=TRUE}, a numeric vector of the point
  estimates.
}
\description{
Make a confidence interval from a point estimate and standard error.
}
