% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaners.R
\name{make_loq}
\alias{make_loq}
\title{Extract numeric, below/above the lower/upper limits of quantification (LLQ
and ULQ), and text data from a vector.}
\usage{
make_loq(
  x,
  llq_pattern = "^< *\\\\(? *([0-9]*\\\\.?[0-9]+) *\\\\)?$",
  ulq_pattern = "^> *\\\\(? *([0-9]*\\\\.?[0-9]+) *\\\\)?$",
  number_pattern = "^(+-)?[0-9]+(\\\\.[0-9]*)?([eE][+-]?[0-9]+)?$",
  replace_llq = 0,
  replace_ulq = Inf,
  ...
)
}
\arguments{
\item{x}{The character vector to extract data from.}

\item{llq_pattern, ulq_pattern}{A regex or vector of regex values for
extracting llq and ULQ.  If a vector, all regexes will be tested.}

\item{number_pattern}{A regex or vector of regex values for finding numbers.}

\item{replace_llq, replace_ulq}{The value to place in the "number" column when
a llq/ulq is matched (typically \code{0} and \code{Inf}, respecitvely or
\code{NA_real_} for both).}

\item{...}{Parameters passed to \code{grep} and \code{grepl} for
\code{llq_pattern} searching.}
}
\value{
a data.frame with columns named "text", "number", "llq", and "ulq".
}
\description{
Extract numeric, below/above the lower/upper limits of quantification (LLQ
and ULQ), and text data from a vector.
}
\details{
If \code{llq_pattern} or \code{ulq_pattern} are regular expressions
  that extract a value, then the extracted value will be used in the
  "llq"/"ulq" column.  If not, then the value \code{-1} will be inserted.  If
  all "llq"/"ulq" values match the \code{number_pattern}, then the
  "llq"/"ulq" columns will be converted to numeric (if not, a warning will be
  given).
}
\examples{
make_loq(c("1", "A", "<1", ">60"))
make_loq(c("1", "A", "<1", ">60"), replace_llq=NA_real_)
make_loq(c("1", "A", "<1", ">60"), replace_llq=NA_real_, replace_ulq=NA_real_)
make_loq(c("1", "A", "<1", ">ULQ"), replace_llq=NA_real_, replace_ulq=NA_real_)
make_loq(c("1", "A", "<1", ">ULQ"), replace_llq=NA_real_, replace_ulq=NA_real_, ulq_pattern=">ULQ")
}
