% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tab_models_tested.R
\name{make_tab_models_tested}
\alias{make_tab_models_tested}
\title{Create a tibble of tested models}
\usage{
make_tab_models_tested(models, caption)
}
\arguments{
\item{models}{A named list of fitted model objects (that can be run through
the `AIC()` function)}

\item{caption}{The caption attribute to add to the tibble (enabling
`pander::pander()` to automatically add the correct caption)}
}
\value{
A tibble with a caption attribute.  The tibble will have columns of
  "Description", "AIC", and "dAIC".
}
\description{
Create a tibble of tested models
}
