% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_character.R
\name{median_character}
\alias{median_character}
\title{Return the median of a character vector}
\usage{
median_character(x, na.rm = TRUE, tie = c("lower", "upper"), ...)
}
\arguments{
\item{x}{an object for which a method has been defined, or a
    numeric vector containing the values whose median is to be computed.}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}

\item{tie}{Are ties broken using the upper or lower value?}

\item{...}{Passed to `sort()`}
}
\value{
The median value of the vector
}
\description{
Return the median of a character vector
}
\details{
As half-way between two values in  a character vector is not
  defined, in the case of a median that would be between two values, either
  the lower or upper value is selected.
}
