% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindiff.R
\name{mindiff}
\alias{mindiff}
\title{Find the minimum difference between a vector and a set of choices.}
\usage{
mindiff(x, choices, tie = c("first", "last", "median-first", "median-last"))
}
\arguments{
\item{x}{The vector of values to adjust}

\item{choices}{The vector of values which should be matched}

\item{tie}{How should a tie (two choices are equidistant) be managed? (See
details.)}
}
\value{
The vector of \code{x} replaced by the closest values from the
  \code{choices} vector.
}
\description{
Find the minimum difference between a vector and a set of choices.
}
\seealso{
\code{\link{mindiff_after}}
}
