% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\docType{data}
\name{NA_logLik_}
\alias{NA_logLik_}
\title{An NA-equivalent value for logLik}
\format{
An object of class \code{logLik} of length 1.
}
\usage{
NA_logLik_
}
\description{
An NA-equivalent value for logLik
}
\seealso{
\code{\link{logLik.xpose_data}}
}
\keyword{datasets}
