% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_missing.R
\name{paste_missing}
\alias{paste_missing}
\title{Concatenate strings dropping missing values}
\usage{
paste_missing(
  ...,
  sep = " ",
  collapse = NULL,
  missing_values = NA,
  paste_last = FALSE
)
}
\arguments{
\item{..., sep, collapse}{See \code{?paste}}

\item{missing_values}{Values considered missing to be ignored in pasting.}

\item{paste_last}{When all \code{...} arguments have been combined and only
one remains, should \code{paste} be called on that last argument?  (Ignored
if \code{collapse} is not \code{NULL}.)}
}
\value{
A character vector of pasted values.
}
\description{
Concatenate strings dropping missing values
}
\details{
If all values are missing, the value from the first argument is
  preserved.  \code{paste_last} affects the final output; the main difference
  is that if \code{FALSE}, \code{NA_character_} values will be preserved, and
  if \code{TRUE}, \code{NA_character_} values will be converted to "NA" (as
  is the case with \code{paste()}).
}
