% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_data.R
\name{patch_data}
\alias{patch_data}
\title{Patch one data set using another (deprecated)}
\usage{
patch_data(...)
}
\arguments{
\item{...}{Ignored}
}
\description{
Patch one data set using another (deprecated)
}
