% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_combine_multi.R
\name{pdf_combine_multi}
\alias{pdf_combine_multi}
\title{Combine multiple PDF files including optional selection of pages.}
\usage{
pdf_combine_multi(input, pages, output = NULL, password = "")
}
\arguments{
\item{input}{path or url to the input pdf file}

\item{pages}{a vector with page numbers to rotate}

\item{output}{base path of the output file(s)}

\item{password}{string with password to open pdf file}
}
\description{
Combine multiple PDF files including optional selection of pages.
}
