% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realize_addl_single.R
\name{realize_addl_single}
\alias{realize_addl_single}
\title{Expand II and ADDL with NONMEM-like methods (removed)}
\usage{
realize_addl_single(...)
}
\arguments{
\item{...}{Ignored}
}
\description{
Expand II and ADDL with NONMEM-like methods (removed)
}
