% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{remove_ggplot_legend}
\alias{remove_ggplot_legend}
\title{Remove the legend(s) from a ggplot}
\usage{
remove_ggplot_legend(x)
}
\arguments{
\item{x}{A ggplot2 object}
}
\value{
The object with legends hidden.
}
\description{
Remove the legend(s) from a ggplot
}
\seealso{
Other plot legend helpers:
\code{\link[=extract_ggplot_legend]{extract_ggplot_legend()}},
\code{\link[=plot_grid_one_legend]{plot_grid_one_legend()}}
}
\concept{plot legend helpers}
