% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonym.R
\name{replace_synonym_list}
\alias{replace_synonym_list}
\title{Use a regular expression to choose the synonyms in a list to replace in a data.frame.}
\usage{
replace_synonym_list(
  x,
  synonyms,
  pattern = "^synonym",
  pattern_ignore_case = TRUE,
  ...
)
}
\arguments{
\item{x}{The values to possibly replace}

\item{synonyms}{A named character vector where the names are the verbatim
value and the values are the values to use for replacement.  When `x` is a
data.frame, this may be a data.frame with columns named
`replacement_column`, `verbatim_column`, and `preferred_column` or a list
of such data.frames.}

\item{pattern}{The regular expression to use to choose the names from
`synonyms` to use for replacement.}

\item{pattern_ignore_case}{Should case be ignored for name matching in
`synonyms`?}

\item{...}{Passed to `replace_synonym()`}
}
\value{
The output of `replace_synonym()` applied to a subset of the list,
  `synonyms`.
}
\description{
Use a regular expression to choose the synonyms in a list to replace in a data.frame.
}
\seealso{
Other Synonyms:
\code{\link[=replace_synonym]{replace_synonym()}}
}
\concept{Synonyms}
