% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaners.R
\name{round_to_fraction}
\alias{round_to_fraction}
\title{Round a value to a fraction, optionally then rounding to a number of digits.}
\usage{
round_to_fraction(x, denominator, digits = Inf)
}
\arguments{
\item{x}{The number to round}

\item{denominator}{The denominator of the fraction for rounding}

\item{digits}{The number of digits for rounding after rounding to the
fraction (\code{Inf} indicates no subsequent rounding)}
}
\value{
x rounded to a decimal value that has an integer numerator relative
  to \code{denominator} (possibly subsequently rounded to a number of decimal
  digits).
}
\description{
Round a value to a fraction, optionally then rounding to a number of digits.
}
\examples{
round_to_fraction(0.7, 3)
}
