% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_exp.R
\name{seq_exp}
\alias{seq_exp}
\title{Sequence generation on an expoential scale}
\usage{
seq_exp(from = 1, to = 1, length.out = NULL, by = NULL,
  add_values = double(length = 0), ...)
}
\arguments{
\item{from}{the starting and (maximal) end values of the
    sequence.  Of length \code{1} unless just \code{from} is supplied as
    an unnamed argument.}

\item{to}{the starting and (maximal) end values of the
    sequence.  Of length \code{1} unless just \code{from} is supplied as
    an unnamed argument.}

\item{length.out}{desired length of the sequence.  A
    non-negative number, which for \code{seq} and \code{seq.int} will be
    rounded up if fractional.}

\item{by}{Explicitly ignored}

\item{add_values}{Add more values to the sequence}

\item{...}{Passed to `base::seq()`}
}
\value{
A sorted sequence
}
\description{
Sequence generation on an expoential scale
}
\details{
If `from` is zero, then it will be modified to be
  `to/(100*length.out)` and zero will be added to the `add_values` vector.
}
