% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_baseline.R
\name{set_baseline}
\alias{set_baseline}
\title{Set a single baseline value given input data.}
\usage{
set_baseline(x, time, min_bl_time = -Inf, max_bl_time = 0,
  summaryfun = mean, ...)
}
\arguments{
\item{x}{A vector of observations}

\item{time}{A vector of times when the observations occurred}

\item{min_bl_time, max_bl_time}{The minum and maximum possible times for
baseline measurements.}

\item{summaryfun}{What functions should summarize or modify the baseline
value(s)?}

\item{...}{Passed to `summaryfun()`}
}
\value{
A scalar value of the baseline measurement
}
\description{
Set a single baseline value given input data.
}
\details{
Possible baseline values are selected when:
\itemize{
  \item{`x` is not `NA`}
  \item{`time` is not `NA`}
  \item{`min_bl_time <= time`}
  \item{`time <= max_bl_time`}
}

If no value is possible, NA is returned (of the same class as `x`).  Within
the possible baseline values, the values with the maximum time are selected,
and all selected values are summarized with `summaryfun(selected_values, ...)`
}
