% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaners.R
\name{setdiff_bidir}
\alias{setdiff_bidir}
\title{Bidirectional setdiff rather than the default unidirectional version}
\usage{
setdiff_bidir(x, y)
}
\arguments{
\item{x, y}{As in \code{setdiff}}
}
\value{
A vector of the two items with setdiff run in both 
  directions.  Items are identified from the original set by names
  starting with "x" and "y".
}
\description{
Bidirectional setdiff rather than the default unidirectional version
}
\examples{
setdiff_bidir(1:5, 3:8)
}
