% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprintf_data_frame.R
\name{sprintf_data_frame}
\alias{sprintf_data_frame}
\alias{sprintf_data_frame_single}
\title{Create new columns in a data.frame with sprintf results}
\usage{
sprintf_data_frame(data, ...)

sprintf_data_frame_single(data, format)
}
\arguments{
\item{data}{the data to use for formatting}

\item{...}{a named list of character vectors.  Names are new columns for
\code{data}, and values are sent to \code{format} in
\code{sprintf_data_frame_single}.}

\item{format}{A named character vector where the names are column names in
\code{data} and the values are sprintf format strings for the column.}
}
\value{
The data frame with columns added for the names of \code{...}.

A character vector with one element per row of \code{data}.
}
\description{
Create new columns in a data.frame with sprintf results
}
\section{Functions}{
\itemize{
\item \code{sprintf_data_frame_single}: Generate a character vector based on sprintf
input formats
}}

\examples{
sprintf_data_frame(
  data=mtcars,
  cyl_mpg=c(mpg="\%g miles/gallon, ", cyl="\%g cylinders"),
  disp_hp=c(disp="\%g cu.in. displacement, ", hp="\%g hp")
)
sprintf_data_frame_single(
  data=mtcars,
  format=c(mpg="\%g miles/gallon, ", cyl="\%g cylinders")
)
}
