% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_value.R
\name{translate_value}
\alias{translate_value}
\alias{translate_value.data.frame}
\alias{translate_value.factor}
\alias{translate_value.character}
\alias{translate_value.default}
\title{Convert a value from something old to something new (using gsub)}
\usage{
translate_value(x, old, new, ...)

\method{translate_value}{data.frame}(x, old, new, ...,
  exclude_col = NULL)

\method{translate_value}{factor}(x, old, new, ...)

\method{translate_value}{character}(x, old, new, ..., fixed = TRUE)

\method{translate_value}{default}(x, old, new, ...)
}
\arguments{
\item{x}{The object to translate}

\item{old, new}{The old and new character strings}

\item{...}{Passed to gsub}

\item{exclude_col}{A vector of columns to exclude from data.frame translation}

\item{fixed}{Passed to gsub}
}
\value{
The object with old converted to new
}
\description{
Convert a value from something old to something new (using gsub)
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Convert all columns (unless excluded)

\item \code{factor}: Convert the levels

\item \code{character}: Use gsub.

\item \code{default}: No translation done.
}}

