% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicode_to_ascii.R
\name{unicode_to_ascii}
\alias{unicode_to_ascii}
\alias{unicode_to_ascii.default}
\alias{unicode_to_ascii.character}
\alias{unicode_to_ascii.factor}
\alias{unicode_to_ascii.logical}
\alias{unicode_to_ascii.NULL}
\alias{unicode_to_ascii.numeric}
\alias{unicode_to_ascii.data.frame}
\alias{unicode_to_ascii.list}
\title{Convert from Unicode to ASCII}
\usage{
unicode_to_ascii(x, ...)

\method{unicode_to_ascii}{default}(x, verbose = FALSE, ...)

\method{unicode_to_ascii}{character}(
  x,
  verbose = FALSE,
  pattern = c("μ", "µ"),
  replacement = c("u", "u"),
  general_conversion = TRUE,
  ...
)

\method{unicode_to_ascii}{factor}(x, ...)

\method{unicode_to_ascii}{logical}(x, verbose = FALSE, ...)

\method{unicode_to_ascii}{`NULL`}(x, verbose = FALSE, ...)

\method{unicode_to_ascii}{numeric}(x, verbose = FALSE, ...)

\method{unicode_to_ascii}{data.frame}(x, verbose = FALSE, ...)

\method{unicode_to_ascii}{list}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{Object that may contain character strings for conversion.}

\item{verbose}{Provide messages about number of values changed.}

\item{pattern, replacement, ...}{Passed to \code{stringi::stri_replace_all_fixed()}}

\item{general_conversion}{After the initial, specific conversions, run
\code{stringi::stri_trans_general(x,
id="Any-Latin;Greek-Latin;Latin-ASCII")} on the current strings.}
}
\value{
An object of the same class as \code{x} with Unicode characters
  converted to ASCII.
}
\description{
Convert from Unicode to ASCII
}
