% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_same.R
\name{coalesce_same}
\alias{coalesce_same}
\alias{coalesce_same.default}
\alias{coalesce_same.data.frame}
\title{Replace missing values with non-missing while confirming that non-missing
values are the same.}
\usage{
coalesce_same(...)

\method{coalesce_same}{default}(...)

\method{coalesce_same}{data.frame}(...)
}
\arguments{
\item{...}{Values to fill in and compare}
}
\value{
The value filled in
}
\description{
Replace missing values with non-missing while confirming that non-missing
values are the same.
}
\details{
For data.frames, combines columns that match names from the first
  data.frame (extra columns from subsequent data.frames are ignored).
}
\seealso{
`dplyr::coalesce`
}
