% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_helpers.R
\name{cumsum_reset}
\alias{cumsum_reset}
\title{Perform a cumulative sum resetting to zero whenever a \code{reset} value is
found.}
\usage{
cumsum_reset(x, reset = NA)
}
\arguments{
\item{x}{The vector to sum over}
}
\value{
A vector the same length as \code{x}
}
\description{
Perform a cumulative sum resetting to zero whenever a \code{reset} value is
found.
}
