% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{extract_ggplot_legend}
\alias{extract_ggplot_legend}
\title{Extract the legend from a ggplot2 object}
\usage{
extract_ggplot_legend(x)
}
\arguments{
\item{x}{A ggplot2 object}
}
\value{
The legend grob from that object
}
\description{
Extract the legend from a ggplot2 object
}
\seealso{
Other plot legend helpers: \code{\link{plot_grid_one_legend}},
  \code{\link{remove_ggplot_legend}}
}
\concept{plot legend helpers}
