% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sub_table.R
\name{find_sub_table}
\alias{find_sub_table}
\title{When data are provided as blocks of tables within a larger data set, 
extract those sub-tables as a list.}
\usage{
find_sub_table(data, value_search, edge_search, ...)
}
\arguments{
\item{data}{The data to extract from}

\item{value_search}{The data value to search for or a function to use
for the search.  If a function, it should take in the data plus the
\code{...} arguments and return a two-column data frame of
\code{row} and \code{column} indices to each of the values.}

\item{edge_search}{A named list with either functions to search for 
the edge of the table or an integer indicating how far to move in 
that direction to find the edge.  If a function, the function 
should take in the data, a row index (integer), a column index 
(integer) to start the search, a direction (one of the values in 
\code{search_order}), and a list indicating \code{found_edges}. The
function should return a single integer of the number of cells in 
the given direction.  The names for the list elements must be 
"left", "right", "up", and "down", and the order of the list 
elements defines the order that they will be searched.}

\item{...}{Arguments passed to user-provided search functions.}
}
\value{
A list of data frames (or similar) with the found sub-tables.
}
\description{
When data are provided as blocks of tables within a larger data set, 
extract those sub-tables as a list.
}
\seealso{
Other Sub-table finding: \code{\link{search_fun_edge}},
  \code{\link{search_fun_values_or_edge}}
}
\concept{Sub-table finding}
