% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_dtc.R
\name{impute_dtc}
\alias{impute_dtc}
\title{Impute dates and times when data are missing.}
\usage{
impute_dtc(data)
}
\arguments{
\item{data}{A data.frame or equivalent object with at least the columns
defined in the details section.}
}
\value{
`data` with the columns "ADTC_IMPUTE_METHOD" and "ADTC_IMPUTED"
  added.
}
\description{
Dates and times will be imputed based on the following rules:
}
\details{
* If both date and time are observed, no the observed value will be used.
* Data are assumed to be grouped by appropriate grouping factors within a
  nominal time so that all times may be at the same time, and data are
  assumed to be sorted in the order specified in the protocol.
* If nominal time since first dose (NTSFD) is missing, no imputation will be
  performed (the measure is assumed to be unscheduled).
* If dates differ within a nominal time measurement, no imputation will be
  performed (a data issue would appear to exist in that case).
* If only one date exists within a nominal time measurement, missing dates
  will be assumed to match the observed date.
* If only one time exists within a nominal interval, all measurements in the
  interval will be assigned to the time that exists.
* If more than one time exists within the nominal interval:
    * Any measurements before the first time that exists will be assigned to
      the first time.
    * Any measurements after the last time that exists will be assigned to
      the last time.
    * Any measurements between the first and last time will be imputed as
      half-way between the measurements that do exist.

Columns used in calculation are:

* ADTC: (the date and time) formatted as an ISO8601 datetime without the time
  zone (yyyy-mm-ddThh:mm:ss) where the entire time or the seconds parts are
  optional.
* STUDYID, USUBJID, NTSFD: grouping variables for the study number, subject
  identifier, and nominal time since first dose.
}
