% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_control.R
\name{join_control}
\alias{join_control}
\alias{join_many_to_one}
\alias{join_one_to_one}
\title{Perform join where the outcome of the join is verifed to match an expected
pattern.}
\usage{
join_control(x, y, join_fun, x_control = "any", y_control = "any", ...)

join_many_to_one(x, y)

join_one_to_one(x, y)
}
\arguments{
\item{x, y}{tbls to join}

\item{join_fun}{Any function that can combine x and y (called as
\code{join_fun(x, y, ...)}).  Typically this will be one of \code{dplyr::left_join},
\code{dplyr::right_join}, etc.}

\item{x_control, y_control}{What outcome is expected from the \code{x}, and \code{y} tbls? Default is "any" (see details).}

\item{...}{Passed to \code{join_fun()}}
}
\value{
A joined tbl
}
\description{
Perform join where the outcome of the join is verifed to match an expected
pattern.
}
\details{
Options for \code{x_control} and \code{y_control} are below and may be
  combined:

\itemize{
\item{\code{"any"}: Any outcome is acceptable; this overrides all other options.}
\item{\code{"all"}: Each row from the input must appear in the output at least one
  time.}
\item{\code{"unique"}: A row may appear in the output zero or one time.}
\item{\code{"missing"}: At least one row must not match in the new dataset (the values
  must be missing).  This option is rarely used.}
\item{\code{"nomissing"}: All rows must match in the new dataset (the values must not
  be missing).}
}

The combination of \code{x_control=c("all", "unique", "nomissing")} (or
\code{y_control}) is a common need to confirm that all values are present exactly
one time and that there are no missing values.
}
\section{Functions}{
\itemize{
\item \code{join_many_to_one}: For the common task of many-to-one mapping, the
helper function `join_many_to_one()` works.

\item \code{join_one_to_one}: For the common task of one-to-one mapping, the
helper function `join_one_to_one()` works.
}}

