% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.gg.R
\name{knit_print.gg}
\alias{knit_print.gg}
\title{Print a ggplot object with space around it}
\usage{
\method{knit_print}{gg}(
  x,
  ...,
  fig_prefix,
  fig_suffix,
  filename = NULL,
  width = 6,
  height = 4,
  units = "in"
)
}
\arguments{
\item{x}{The plot object}

\item{...}{Passed to `print`.}

\item{fig_prefix}{See `fig_suffix`}

\item{fig_suffix}{Character strings passed to `cat` before and after printing
`x` (if not missing).}

\item{filename}{Save the figure to the filename, if provided}

\item{width, height, units}{passed to `ggplot2::ggsave()`}
}
\value{
`x` invisibly
}
\description{
Print a ggplot object with space around it
}
\seealso{
\code{\link{knit_print.gg_list}}
}
