% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.gg.R
\name{knit_print.gg_list}
\alias{knit_print.gg_list}
\title{Print a list of ggplot objects with space around each}
\usage{
\method{knit_print}{gg_list}(x, ..., filename = NULL, fig_suffix = "\\n\\n")
}
\arguments{
\item{x}{A list of plot objects}

\item{...}{Passed to \code{knit_print}.}

\item{filename}{Save the figure to the filename, if provided.  If the
filename contains "%d" (optionally with sprintf-formatting instructions
such as "%03d"), then the filename will be run through
\code{sprintf(filename, seq_along(x))} to generate the filename.}

\item{fig_suffix}{Character strings passed to `cat` before and after printing
`x` (if not missing).}
}
\value{
\code{x} invisibly
}
\description{
Print a list of ggplot objects with space around each
}
\seealso{
`[knit_print.gg]`
}
