% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_helpers.R
\name{latex_reference}
\alias{latex_reference}
\alias{latex_label}
\alias{latex_label_clean}
\alias{latex_label_first_last}
\title{Generate a latex label or reference to the label}
\usage{
latex_reference(x)

latex_label(x)

latex_label_clean(x)

latex_label_first_last(
  object,
  text,
  sep = "-",
  suffix_text = c("first", "last")
)
}
\arguments{
\item{x}{The label (a character vector, usually starting with "fig:" for figures and "tab:" for tables)}

\item{object}{The object (with a length > 1)}

\item{text}{The text to use for the start of the label (usually "fig:xxx" or
"tab:xxx")}

\item{sep}{The separator to use between the text and the suffix}

\item{suffix_text}{a 2-long character string with the text to use after the
separator}
}
\value{
A knitr asis_output label or reference

A character vector with latex labels for the first and last objects
  and blank strings in between.
}
\description{
Generate a latex label or reference to the label
}
\section{Functions}{
\itemize{
\item \code{latex_label}: Generate a latex label (location to be found by a reference).

\item \code{latex_label_clean}: Clean latex labels so that they are usable
(remove spaces and pass through \code{Hmisc::latexTranslate}).

\item \code{latex_label_first_last}: Generate a latex label for the first and last of
a list of objects with blanks between
}}

