% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_helpers.R
\name{latex_reference}
\alias{latex_reference}
\alias{latex_label}
\alias{latex_label_clean}
\title{Generate a latex label or reference to the label}
\usage{
latex_reference(x)

latex_label(x)

latex_label_clean(x)
}
\arguments{
\item{x}{The label (a character vector, usually starting with "fig:" for figures and "tab:" for tables)}
}
\value{
A knitr asis_output label or reference
}
\description{
Generate a latex label or reference to the label
}
\section{Functions}{
\itemize{
\item \code{latex_label}: Generate a latex label (location to be found by a reference).

\item \code{latex_label_clean}: Clean latex labels so that they are usable
(remove spaces and pass through \code{Hmisc::latexTranslate}).
}}

