% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_helpers.R
\name{make_ci}
\alias{make_ci}
\title{Make a confidence interval from a point estimate and standard error.}
\usage{
make_ci(point, se, level = 0.95, transform = NULL, numeric = FALSE,
  format_numeric = "\%0.3g", format_character = "\%s")
}
\arguments{
\item{point}{The point estimate (numeric vector)}

\item{se}{The standard error of the estimate (numeric vector)}

\item{level}{The confidence level}

\item{transform}{A function that takes a matrix input and returns a matrix
output of the confidence interval values transformed.}

\item{numeric}{Should the transformed point estimate be returned instead of
the character representation?}

\item{format_numeric, format_character}{Formats for use in \code{sprintf} for
the output of the \code{transform} function.}
}
\value{
If \code{numeric=FALSE} a character vector of the confidence
  intervals represented as "X [X, X]" where X are numbers with three
  significant figures (or if the transform returns a character matrix, those
  text).  If \code{numeric=TRUE}, a numeric vector of the point estimates.
}
\description{
Make a confidence interval from a point estimate and standard error.
}
