% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manage.R
\name{nonmem_column_order}
\alias{nonmem_column_order}
\title{Output a data.frame with numeric columns on the left.}
\usage{
nonmem_column_order(x)
}
\arguments{
\item{x}{A data.frame or similar object}
}
\description{
Output a data.frame with numeric columns on the left.
}
\seealso{
Other Data Management: 
\code{\link{check_expected_cols}()},
\code{\link{get_data_manage_standard_cols}()}
}
\concept{Data Management}
