% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_tools.R
\docType{data}
\name{number_patterns}
\alias{number_patterns}
\title{Regular expression patterns for numbers}
\format{An object of class \code{list} of length 8.}
\usage{
number_patterns
}
\description{
Regular expression patterns for numbers
}
\details{
\itemize{
  \item{natural}{positive integers without a sign}
  \item{nonnegative_integer_no_sign}{only digits}
  \item{integer}{an optional +- at the beginning followed by only digits
    (note that "-0" is valid)}
  \item{float_no_sign}{digits, a decimal point, and more digits}
  \item{float_or_integer_relaxed}{an optional +- followed by one of the
    following:}
  \itemize{
    \item{digits}
    \item{digits then a decimal point}
    \item{a decimal point then digits}
    \item{digits then a decimal point then digits}
  }
  \item{float_or_integer_strict}{an optional +- followed by digits optionally
    followed by a decimal point and digits (starting or ending with a decimal
    point is not allowed).}
  \item{scientific_notation}{Strict scientific notation made of the
    following, in order:}
  \itemize{
    \item{an optional +-}
    \item{a coefficient that is zero (with an optional decimal point and
      optional additinoal zeros), a natural number, or a float that does not
      start with zero}
    \item{one of "e", "E", "d", or "D"}
    \item{a optional +-}
    \item{integer}
  }
  \item{scientific_notation_relaxed}{Relaxed scientific notation (relaxing on
    the coefficient rules) made of the following, in order:}
  \itemize{
    \item{float_or_integer_relaxed}
    \item{one of "e", "E", "d", or "D"}
    \item{a optional +-}
    \item{integer}
  }
}
}
\examples{
grepl(pattern=number_patterns$scientific_notation_relaxed, "1e5")
grepl(pattern=number_patterns$scientific_notation_relaxed, "1e-5")
grepl(pattern=number_patterns$scientific_notation_relaxed, "1e.5")
}
\keyword{datasets}
