% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plot_grid_one_legend}
\alias{plot_grid_one_legend}
\title{Extract the legend from the first figure, remove the legends from all plots,
and put the legend at the end.}
\usage{
plot_grid_one_legend(...)
}
\arguments{
\item{...}{One or more ggplot2 objects}
}
\value{
A list of the ggplot2 objects (suitable for the \code{plotlist}
  argument of \code{cowplot::plot_grid})
}
\description{
Extract the legend from the first figure, remove the legends from all plots,
and put the legend at the end.
}
\seealso{
Other plot legend helpers: \code{\link{extract_ggplot_legend}},
  \code{\link{remove_ggplot_legend}}
}
\concept{plot legend helpers}
