% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realize_addl_single.R
\name{realize_addl_single}
\alias{realize_addl_single}
\title{Expand II and ADDL with NONMEM-like methods for a single subject}
\usage{
realize_addl_single(time, evid, addl = 0, ii = 0)
}
\arguments{
\item{time}{Time of assessment}

\item{evid}{NONMEM-style event identifier; 1 is a dose; 4 is a reset and dose; all other rows are ignored.}

\item{addl}{The number of additional doses (0 indicates one dose will be administered)}

\item{ii}{The interdose-interval (time between doses)}
}
\value{
A data.frame with rows expanded based on addl and ii.  The final
  number of rows will be \code{sum(evid %in% c(1, 4)) + sum(addl[evid %in%
  c(1, 4)])}
}
\description{
Expand II and ADDL with NONMEM-like methods for a single subject
}
