% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_synonym.R
\name{replace_synonym}
\alias{replace_synonym}
\title{Replace values in a character string with their synonym}
\usage{
replace_synonym(x, synonyms, ignore_case = TRUE)
}
\arguments{
\item{x}{The values to possibly replace}

\item{synonyms}{A named character vector where the names are the verbatim value and the values are the values to use for replacement.}

\item{ignore_case}{Should the synonyms be replaced case-insensitively?}
}
\value{
`x` with preferred values instead of verbatim values.
}
\description{
Replace values in a character string with their synonym
}
\examples{
replace_synonym(
  c("A", "B", "C", "a"),
   c(A="apple", B="bear", C="cabbage")
)
replace_synonym(
  c("A", "B", "C", "a"),
  c(A="apple", B="bear", C="cabbage"),
  ignore_case=FALSE
)
}
