% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sub_table.R
\name{search_fun_edge}
\alias{search_fun_edge}
\title{Go to the edge}
\usage{
search_fun_edge(data, row, column, direction, found_edges, ...)
}
\arguments{
\item{data, row, column, direction, found_edges}{See \code{\link{find_sub_table}}}

\item{...}{Ignored}
}
\value{
A number for the distance to the edge.
}
\description{
Go to the edge
}
\seealso{
Other Sub-table finding: 
\code{\link{find_sub_table}()},
\code{\link{search_fun_values_or_edge}()}
}
\concept{Sub-table finding}
