% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sub_table.R
\name{search_fun_values_or_edge}
\alias{search_fun_values_or_edge}
\title{Go to a value or the edge}
\usage{
search_fun_values_or_edge(values, from = c("row", "column"),
  skip = c(0, 0), exclude_value = FALSE)
}
\arguments{
\item{values}{The value to search for}

\item{from}{The location to search from within the block.  It must 
already be found.  Options pairs of "up", "row", or "down" and
"left", "column", or "right".  Defaults to the column and row of
the found value.}

\item{skip}{The rows and columns to skip from the \code{from} 
argument when starting the search.}

\item{exclude_value}{Exclude the found value from the edge?  (Useful
if you are searching for an NA cell at the edge and do not want
that NA value in the result.)}
}
\value{
A number for the distance to the edge.
}
\description{
Go to a value or the edge
}
\seealso{
Other Sub-table finding: \code{\link{find_sub_table}},
  \code{\link{search_fun_edge}}
}
\concept{Sub-table finding}
