% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprintf_data_frame.R
\name{sprintf_data_frame}
\alias{sprintf_data_frame}
\alias{sprintf_data_frame_single}
\title{Create new columns in a data.frame with sprintf results}
\usage{
sprintf_data_frame(data, ..., factor_out_if_factor_in = TRUE, ordered = NULL)

sprintf_data_frame_single(
  data,
  format,
  factor_out_if_factor_in = TRUE,
  ordered = NULL
)
}
\arguments{
\item{data}{the data to use for formatting}

\item{...}{a named list of character vectors.  Names are new columns for
\code{data}, and values are sent to \code{format} in
\code{sprintf_data_frame_single}.}

\item{factor_out_if_factor_in}{If any of the input columns are factors, make
the output column a factor in the same order as the input column factors}

\item{ordered}{If \code{factor_out_if_factor_in} converts the output to a
factor, pass to \code{base::factor}.  If \code{NULL}, then it is set to
\code{TRUE} if any of the input columns are ordered factors.}

\item{format}{A named character vector where the names are column names in
\code{data} and the values are sprintf format strings for the column.}
}
\value{
The data frame with columns added for the names of \code{...}.

A character vector with one element per row of \code{data}.
}
\description{
Create new columns in a data.frame with sprintf results
}
\section{Functions}{
\itemize{
\item \code{sprintf_data_frame_single()}: Generate a character vector based on sprintf
input formats

}}
\examples{
sprintf_data_frame(
  data=mtcars,
  cyl_mpg=c(mpg="\%g miles/gallon, ", cyl="\%g cylinders"),
  disp_hp=c(disp="\%g cu.in. displacement, ", hp="\%g hp")
)
sprintf_data_frame_single(
  data=mtcars,
  format=c(mpg="\%g miles/gallon, ", cyl="\%g cylinders")
)
}
